requirements.txt
pint
matplotlib
python-dotenv
streamlit
numpy